refactor: Step3,4 검색결과 컴포넌트 분리
