+++ src/components/plan/index.ts
@@ -2,4 +2,6 @@ export { default as CustomStepper } from './CustomStepper.vue';
 export { default as Step1DateSetting } from './Step1DateSetting.vue';
 export { default as Step2FriendInvite } from './Step2FriendInvite.vue';
 export { default as Step3AccommodationSetting } from './Step3AccommodationSetting.vue';
+export { default as Step3AccommodationDrawer } from './Step3AccommodationDrawer.vue';
 export { default as Step4PlaceSearch } from './Step4PlaceSearch.vue';
+export { default as Step4PlaceDrawer } from './Step4PlaceDrawer.vue';
